Simplified raster read functions.
I've removed a number of redundant read functions and augmented the return results of the raster-seq function.

Instead of implementing every overloaded version of the read-raster-direct function, I've decided to start with the simple cases first. Passing in buffer parameters requires a lot of upfront configuration. The read-raster-direct now relies on the ReadRaster_Direct implementation that automatically initializes a buffer. Although this will lead to slower performance, it does simplify things for the time being. This dramatically simplifies the raster-seq function and has the added benefit of it now being parallelization friendly as the buffer is no longer shared!